Keep readable URL helper off play-view root
